+++ src/oihana/files/TarFile.php
@@ -8,6 +8,9 @@ use oihana\enums\Char;
 use oihana\enums\FileExtension;
 use oihana\exceptions\DirectoryException;
 use oihana\exceptions\FileException;
+
+use function oihana\core\files\assertDirectory;
+use function oihana\core\files\assertFile;
 use function oihana\core\files\createDirectory;
 
 /**
@@ -20,21 +23,19 @@ final class TarFile
      * Supported tar‑related extensions (lower‑case, with leading dot).
      * @var string[]
      */
-    public const array TAR_EXTENSIONS = [ FileExtension::GZ , FileExtension::TAR , FileExtension::TGZ , FileExtension::TAR_GZ ] ;
+    public const array TAR_EXTENSIONS = [ FileExtension::GZ , FileExtension::TAR , FileExtension::TGZ , FileExtension::TAR_BZ2 , FileExtension::TAR_GZ ] ;
 
     /**
      * Heuristic: validate a file is a tar or gzip‑compressed tar archive.
      * @param string $file The file to evaluates.
-     * @return bool True if the file is a tar file, false otherwise.
+     * @param array $mimeTypes The mimetypes to validate the file.
+     * @return true True if the file is a tar file, false otherwise.
      * @throws FileException If the archive does not exist or is not a tar file.
      * @throws RuntimeException When extraction fails or directory cannot be created.
      */
-    public function assertFile( string $file ): bool
+    public function assertFile( string $file , array $mimeTypes = [ 'x-tar' , 'gzip' , 'bzip2'] ): true
     {
-        if ( !is_file( $file ) )
-        {
-            throw new FileException( sprintf('The file "%s" does not exist.' , $file ) ) ;
-        }
+        assertFile( $file );
 
         $extension = Char::DOT . strtolower( (string) pathinfo( $file, PATHINFO_EXTENSION ) ) ;
 
@@ -43,16 +44,27 @@ final class TarFile
             throw new FileException( sprintf('File "%s" does not have a tar‑related extension.' , $file ) ) ;
         }
 
-
         $info     = finfo_open(FILEINFO_MIME_TYPE ) ;
         $mimeType = finfo_file( $info , $file ) ?: Char::EMPTY ;
 
         finfo_close( $info ) ;
 
-        if ( !str_contains($mimeType, 'x-tar') && !str_contains( $mimeType, 'gzip' ) )
+        $mimeTypeMatches = false;
+        foreach ($mimeTypes as $type)
         {
-            throw new FileException(sprintf('File "%s" is not recognised as a tar archive (Mime-Type: %s).' , $file , $mimeType ) ) ;
+            if ( str_contains( $mimeType , $type ) )
+            {
+                $mimeTypeMatches = true ;
+                break;
+            }
         }
+
+        if ( !$mimeTypeMatches )
+        {
+            throw new FileException( sprintf('File "%s" is not recognised as a tar archive (Mime-Type: %s).' , $file , $mimeType ) ) ;
+        }
+
+        return true;
     }
 
     /**
@@ -65,10 +77,7 @@ final class TarFile
      */
     public function create( string $directory, string $extension = FileExtension::TAR_GZ ): string
     {
-        if ( !is_dir( $directory ) )
-        {
-            throw new DirectoryException( sprintf('Creates a tar file failed, the passed-in directory %s does not exist.' , json_encode( $directory , JSON_UNESCAPED_SLASHES ) ) ) ;
-        }
+        assertDirectory( $directory ) ;
 
         $archiveName = basename( $directory ) ;
         $tarFile     = $directory . $extension;
@@ -86,7 +95,6 @@ final class TarFile
         }
     }
 
-
     /**
      * Extracts a tar file into a directory.
      * @param string $tarFile The tar file to extract.
@@ -98,9 +106,9 @@ final class TarFile
      */
     public function extract( string $tarFile , string $directory ): true
     {
-        createDirectory( $directory ) ;
-
         $this->assertFile( $tarFile ) ;
+
+        createDirectory( $directory ) ;
 
         $command = "tar -xzf $tarFile -C $directory" ;
 
@@ -112,7 +120,16 @@ final class TarFile
         }
         else
         {
-            throw new RuntimeException( sprintf('Extract tar file failed, impossible to extract the tar file %s to directory %s.' , json_encode( $tarFile , JSON_UNESCAPED_SLASHES ) , json_encode( $directory , JSON_UNESCAPED_SLASHES ) ) , $status ) ;
+            throw new RuntimeException
+            (
+                sprintf
+                (
+                    'Extract tar file failed, impossible to extract the tar file %s to directory %s.' ,
+                    json_encode( $tarFile   , JSON_UNESCAPED_SLASHES ) ,
+                    json_encode( $directory , JSON_UNESCAPED_SLASHES )
+                )
+                , $status
+            ) ;
         }
     }
 }